Fixed JSON Parse Error

+++ Cloudy.App/cloudy/lib/models/weather_data.dart
@@ -190,7 +190,7 @@ class Main {
 }
 
 class Wind {
-  int? speed;
+  double? speed;
   int? deg;
 
   @override
